+++ test/IntegrationTests/AspNetTests.cs
@@ -25,6 +25,8 @@ namespace IntegrationTests;
 
 public class AspNetTests
 {
+    private const string ServiceName = "TestApplication.AspNet.NetFramework";
+
     private readonly Dictionary<string, string> _environmentVariables = new();
 
     public AspNetTests(ITestOutputHelper output)
@@ -57,6 +59,34 @@ public class AspNetTests
         await CallTestApplicationEndpoint(webPort);
 
         collector.AssertExpectations();
+    }
+
+    [Fact]
+    [Trait("Category", "EndToEnd")]
+    [Trait("Containers", "Windows")]
+    public async Task TracesResource()
+    {
+        Assert.True(EnvironmentTools.IsWindowsAdministrator(), "This test requires Windows Administrator privileges.");
+
+        _environmentVariables["OTEL_SERVICE_NAME"] = ServiceName;
+
+        // Using "*" as host requires Administrator. This is needed to make the mock collector endpoint
+        // accessible to the Windows docker container where the test application is executed by binding
+        // the endpoint to all network interfaces. In order to do that it is necessary to open the port
+        // on the firewall.
+        using var collector = new MockSpansCollector(Output, host: "*");
+        using var fwPort = FirewallHelper.OpenWinPort(collector.Port, Output);
+        collector.ResourceExpector.Expect("service.name", ServiceName); // this is set via env var in Dockerfile and Wep.config, but env var has precedence
+        collector.ResourceExpector.Expect("deployment.environment", "test"); // this is set via Wep.config
+
+        string collectorUrl = $"http://{DockerNetworkHelper.IntegrationTestsGateway}:{collector.Port}";
+        _environmentVariables["OTEL_TRACES_EXPORTER"] = "otlp";
+        _environmentVariables["OTEL_EXPORTER_OTLP_ENDPOINT"] = collectorUrl;
+        var webPort = TcpPortProvider.GetOpenPort();
+        await using var container = await StartContainerAsync(webPort);
+        await CallTestApplicationEndpoint(webPort);
+
+        collector.ResourceExpector.AssertExpectations();
     }
 
     [Fact]