+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This beta release is built on top of [OpenTelemetry .NET](https://github.com/ope
 
 ### Added
 
+- Support configuring `OTEL_*` settings using `App.config` and `Web.config`.
 - Add Quartz traces instrumentation.
 
 ### Changed